+++ eslint.config.js
@@ -12,6 +12,7 @@ export default tseslint.config(
       parserOptions: {
         projectService: {
           allowDefaultProject: ['scripts/*.mjs'],
+          maximumDefaultProjectFileMatchCount_THIS_WILL_SLOW_DOWN_LINTING: 16,
         },
         tsconfigRootDir: import.meta.dirname,
       },